move create command into its own file
